refactor: add QrCodeUploadButton to QrCodeReader (#69)
